Fixed a bug where choosing "clean slate" during migration prompts was not clearing all data.
Scenario: Login with a user with values in different month > Log out > Login with a new user with no values > Start with a clean slate > Log out immediately.
Expected outcome: Clear all months. Work with a clean UI.
Bug: If logged out of the new user without navigating to other months, previous user's values were persisting in other months.
Fix: Use the same data from the server and pass it to localStorage. If its empty (which is the current case), then it clears all the local data associated with it as well.